Dani: He creat una vista register per a poder registrar usuaris,  juntament amb un botó a la vista de login per redirigirte a crear un usuari. Marc: He depurat codi, he començat a crear la vista i formulari de crear histories, estic preparant per a importar entitats des de la base de dades. Jhon: he afegit un boton nou de logout i un mètode per utilitzar-ho.
Com a treball grupal, hem migrat el treball a visual studio, per poder treballar tots a l'hora i fer més fàcil el desenvolupament

+++ code/Plantilla/src/main/java/com/example/application/data/repositories/ImatgeRepository.java
@@ -16,5 +16,5 @@ public interface ImatgeRepository extends JpaRepository<Imatge, Integer> {
     List<Imatge> search(@Param("searchTerm") String searchTerm);
 
     @EntityGraph(attributePaths={"profilePicture"})
-    Imatge findWithPropertyPictureAttachedById(Integer id);
+    Imatge findWithPropertyPictureAttachedById(Long id);
 }
